Faz o salvamento da atividade no banco de dados

+++ app/src/main/java/br/edu/utfpr/trabalhoconclusaocurso/data/model/Atividade.kt
@@ -5,7 +5,7 @@ data class Atividade(
     val idUsuario: String,
     val nome: String,
     val dataHora: String, // ISO-8601 "2025-09-03T14:30:00"
-    val duracao: Int,     // em segundos
+    val duracao: Long,     // em segundos
     val distancia: Double,
     val velocidadeMedia: Double,
     val caloriasPerdidas: Double?
